+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 # Unreleased
 
 - Remove request URLs forward slash append
+- Add events to `TransactionWorker` module that dapps can listen to
+- Introduce `aptos.transaction.batch` namespace to handle batch transactions
+- Support `aptos.transaction.batch.forSingleAccount()` to send batch transactions for a single account
+- Label `aptos.batchTransactionsForSingleAccount()` as `deprecated` to prefer using `aptos.transaction.batch.forSingleAccount()`
 
 # 1.4.0 (2024-01-08)